rtcpsu: Fixed MISRAC rule 10.3 violation
This patch fix following violation
"misra_c_2012_rule_10_3_violation:Implicit conversion from
 essential type to different or narrower essential type "signed 32-bit int"

Fixed this violation across the driver by explicitly typecasting at all the
required/intended places.


Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/rtcpsu/src/xrtcpsu_selftest.c
@@ -68,7 +68,7 @@
 ******************************************************************************/
 s32 XRtcPsu_SelfTest(const XRtcPsu *InstancePtr)
 {
-	s32 Status = XST_SUCCESS;
+	s32 Status = (s32)XST_SUCCESS;
 	u32 SafetyCheck;
 
 	/* Assert validates the input arguments */
@@ -85,7 +85,7 @@ s32 XRtcPsu_SelfTest(const XRtcPsu *InstancePtr)
 			XRTC_SFTY_CHK_OFFSET);
 
 	if (SafetyCheck != XRTC_SFTY_CHK_RSTVAL) {
-		Status = XST_FAILURE;
+		Status = (s32)XST_FAILURE;
 	}
 
 	return Status;
